requirements.txt
# Qatar AI Platform — Backend
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
httpx>=0.25.0
slowapi>=0.1.9
apscheduler>=3.10.0
groq>=0.4.0
